+++ packages/backtest/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @zapcli/backtest
 
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @zapcli/core@1.0.5
+
 ## 1.0.4
 
 ### Patch Changes